third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/history-traversal/browsing_context_name_cross_origin-expected.txt

This is a testharness.js-based test.
FAIL Restoring window.name on cross-origin history traversalassert_equals: expected 2 but got 1
Harness: the test ran to completion.
